requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
line-bot-sdk
google-generativeai>=0.8.0
gradio
requests